Refactor out createActions.
